% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geodata.R
\name{download_geodata}
\alias{download_geodata}
\title{Download datasets}
\usage{
download_geodata(dataset, path, unzip = TRUE, available_geodata)
}
\arguments{
\item{dataset}{character string containing the name of an available dataset.
Use \code{get_available_geodata} to find the corresponding names.}

\item{path}{the directory to download files to.}

\item{unzip}{if \code{TRUE}, the downloaded folder is unzipped and the zip folder is
deleted.}

\item{available_geodata}{result of a previous call with \code{get_available_geodata}.
Optional argument.}
}
\description{
\code{download_geodata} downloads datasets from the geo-information platform of
    the Swiss Confederation (\url{https://data.geo.admin.ch/}).
}
\details{
The acquisition and use of data or services is free of charge,
    subject to the provisions on fair use (see \url{https://www.geo.admin.ch/terms-of-use}).
}
\examples{
\dontrun{
# Show all available records for the search term Agglomerations
download_geodata("ch.are.reisezeit-oev")
}

#' @export

}
