% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stac_assets.R
\name{download_stac_assets}
\alias{download_stac_assets}
\title{Download assets from STAC API}
\usage{
download_stac_assets(response, path, type)
}
\arguments{
\item{response}{response object of a \code{get_stac_assets} call to the
STAC API.}

\item{path}{the directory to download files to.}

\item{type}{selects the type of file to download. If not specified, all assets are downloaded.
Use \code{type = "image"} to download GeoTIFF only.}
}
\description{
\code{download_stac_assets} downloads assets from the STAC API on the
    geo-information platform of the Swiss Confederation
    (\url{https://data.geo.admin.ch/api/stac/v0.9/}).
}
\details{
The acquisition and use of data or services is free of charge,
    subject to the provisions on fair use (see \url{https://www.geo.admin.ch/terms-of-use}).
}
\examples{
\dontrun{
# Download the aerial photograph of the old town of Aarau (LV03 coordinates)
call_res <- get_stac_assets(
    collection_id = "ch.swisstopo.swissimage-dop10",
    lon = 645685
    lat = 249287
    )

download_stac_assets(call_res, type = "image")

}

}
