% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_geodata.R
\name{get_available_geodata}
\alias{get_available_geodata}
\title{Display all available datasets}
\usage{
get_available_geodata(include_links = FALSE)
}
\arguments{
\item{include_links}{if \code{TRUE}, links to available resources are included in the results.}
}
\value{
A tibble with the names of the datasets and their retrieval function. The links
    to available resources are included if \code{include_links = TRUE}.
}
\description{
\code{get_available_geodata} displays all datasets available on the geo-information platform
    of the Swiss Confederation (\url{https://data.geo.admin.ch/}).
}
\details{
The acquisition and use of data or services is free of charge,
    subject to the provisions on fair use (see \url{https://www.geo.admin.ch/terms-of-use}).
}
\examples{
# Display all available datasets
get_available_geodata()

}
