% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stac_collections.R
\name{get_stac_collections}
\alias{get_stac_collections}
\title{Display all data retrievable via the STAC API}
\usage{
get_stac_collections(
  api = "https://data.geo.admin.ch/api/stac/v0.9/collections"
)
}
\arguments{
\item{api}{link to query endpoint listing all available datasets.}
}
\value{
A tibble with a set of metadata (name, id, description, spatial and temporal extent)
    about all available geospatial datasets.
}
\description{
\code{get_stac_collections} displays a description of all data provided
    via the Spatial Temporal Asset Catalog (STAC) REST Interface on the
    geo-information platform of the Swiss Confederation
    (\url{https://data.geo.admin.ch/api/stac/v0.9/}).
}
\details{
The acquisition and use of data or services is free of charge,
    subject to the provisions on fair use (see \url{https://www.geo.admin.ch/terms-of-use}).
}
\examples{
# Show all available datasets of the STAC API
get_stac_collections()

}
