% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geodata.R
\name{search_geodata}
\alias{search_geodata}
\title{Search available datasets for pattern matches}
\usage{
search_geodata(
  pattern,
  fuzzy_support = FALSE,
  include_links = FALSE,
  available_geodata
)
}
\arguments{
\item{pattern}{character string containing a regular expression to be matched in
the dataset names on the geo-information platform.}

\item{fuzzy_support}{if \code{TRUE}, approximate matches are included.}

\item{include_links}{if \code{TRUE}, the links to available resources are included
in the results.}

\item{available_geodata}{result of a previous call with \code{get_available_geodata}.
Optional argument.}
}
\value{
A tibble with matching datasets and their retrieval function. The links
    to available resources are included if \code{include_links = TRUE}.
}
\description{
\code{search_geodata} searches for matches to argument \code{pattern} within
    the names of all available datasets on the geo-information platform of
    the Swiss Confederation (\url{https://data.geo.admin.ch/}).
}
\details{
The acquisition and use of data or services is free of charge,
    subject to the provisions on fair use (see \url{https://www.geo.admin.ch/terms-of-use}).
}
\examples{
# Show all available datasets for the search term 'Agglomerations'
search_geodata(pattern = "Agglomeration")

# Show all available datasets of the Federal Office for the Environment (BAFU)
search_geodata(pattern = "BAFU")

}
