% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_metadata.R
\name{show_metadata}
\alias{show_metadata}
\title{View metadata}
\usage{
show_metadata(dataset, available_geodata)
}
\arguments{
\item{dataset}{character string containing the name of an available dataset.
Use \code{get_available_geodata()} to find the corresponding names.}

\item{available_geodata}{result of a previous call with \code{get_available_geodata}.
Optional argument.}
}
\description{
\code{show_metadata} opens the entry of a dataset in the Swiss Geometadata
    Catalogue (\url{https://www.geocat.admin.ch}).
}
\details{
The acquisition and use of data or services is free of charge,
    subject to the provisions on fair use (see \url{https://www.geo.admin.ch/terms-of-use}).
}
\examples{
# Show metadata of the catalog of sports facilities of national importance
show_metadata("ch.baspo.nationales-sportanlagenkonzept")

}
