% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_preview.R
\name{show_preview}
\alias{show_preview}
\title{Preview data}
\usage{
show_preview(dataset, available_geodata)
}
\arguments{
\item{dataset}{character string containing the name of an available dataset.
Use \code{get_available_geodata} to find the corresponding names.}

\item{available_geodata}{result of a previous call with \code{get_available_geodata}.
Optional argument.}
}
\description{
\code{show_preview} opens the dataset on mapping platform of
    the Swiss Confederation (\url{https://map.geo.admin.ch}).
}
\details{
The acquisition and use of data or services is free of charge,
    subject to the provisions on fair use (see \url{https://www.geo.admin.ch/terms-of-use}).
}
\examples{
# Shows the preview of the travel time to 6 major centres by public transport
show_preview("ch.are.reisezeit-oev")

}
