% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swissgd.R
\docType{package}
\name{swissgd}
\alias{swissgd}
\title{\code{swissgd} package}
\description{
Interface to the Geo-Information Platform of the Swiss Confederation
}
\details{
See the README on
\href{https://github.com/zumbov2/swissgd#readme}{GitHub}
}
